Fix: SimpleRenderer: methods for rotating the graph
Must not modify values of the transformation matrix m
during the calculation of the matrix product using m.

Therefore, we have changed the code
to store updated values of m into a temporary matrix,
and then transfer the values from there to m
after the calculation of the matrix product complete.