Add: improve the algorithm controlling angles by mouse dragging
We are using 3-axes rotation algorithm
for controlling camera angles by mouse dragging.

However, near the rotational center,
the 3-axes rotation algorithm gives a little "hanged up" feelings.

Hence, we have changed to apply
simple 2-axes rotation algorithm when the mouse is near the center.

When the mouse is far enough from the center,
it is switched to 3-axes rotation algorithm, as it bas been.