Remove: TempMain: TemporaryPreviewWindow
It has been a "temporary" window
for previewing rendered graphs,
while we have developing/debugging the rendering engine.

By previous commit,
we have implemented the essential event handlers
to the main window of RinearnGraph3D,
and now the main window can play the above role enough.

So the "temporary" window has finished its role,
and we have removed it. Thanks!