Add: new class: SpaceConverter
A class for converting coordinates from the real space to the "scaled space".
In the scaled space, each axis of the graph frame are mapped to the range [-1.0, 1.0].

The above logic have originally implemented in Axis class.
However, we consider that the design of Axis class is not suitable for
the current internal structure of this app.
So have moved some code to this new class, before deleting the Axis class.