Add: abstract class and subclasses of data groups (declaration only)
The "data group" is the unit of data composing of
a group of coordinate values, to be plotted on the graph.

For example, "Data loaded from File-A",
"Data loaded from File-B", "Data of Math Expression f(x,y)",
are different data groups, independent each other.

In Model layer, each data group is handled as
an instance of a class inheriting AbstractDataGroup:

[Inheritance tree]

  AbstractDataGroup
    |
    +- ArrayDataGroup (Data loaded from files)
    |
    +- ExpressionDataGroup
        |
        +- ZxyExpressionDataGroup ("Z(x,y)" form expr.)
        |
        +- XtYtZtExpressionDataGroup ("X(t),Y(t),Z(t)" form expr.)