Add: API: setAsynchronousPlottingEnabled(boolean enabled)
Enables/disables the asynchronous plotting feature.

If this feature is disabled (default),
when you plot the data by calling setData(double[], double[], double[]) method,
the processing flow returns to the called-side after when the plotting has competed.

In the contrast, if this feature is enabled, the processing flow returns from
setData(double[], double[], double[]) method to the caller-side immediately
with storing the data to the buffer.
Then, the plotting will be performed asynchronously
on the other thread, at a suitable timing.

This feature is useful when you want to animate the graph
by calling setData method repeatedly in high frequency.
In such case, with enabling this feature,
you can reduce the load of the main thread,
and also can make the animation smooth.