Add: Model: setArrayDataSeries(ArrayDataSeries[]) method
This new method sets all the data series to be plotted.

At first glance, it seems that we can perform
the completely same operation by using clearArrayDataSeries()
and (multiple calls of) addArrayDataSeries(...) methods.

However, this operation must be performed as an "atomic operation"
in some situation, especially when asynchronous-plotting feature is enabled.
Therefore we added this method.