Refactor: move some elements of MathDataSeries(abstract) to subclasses
Firstly, we defined many fields in this abstract class
to avoid duplicate code.

But now we feel that it was bad way.
It is not readable that the abstract class has many implementational code.
It should be a simple, "abstract concept of subclasses".

So we have moved some fields and their getter methods
from the abstract class, MathDataSeries, to its subclasses.