Add: new class: DataStore
The class storing and managing data to be plotted.

Until this commit,
all the data series to be plotted are directly declared in Model class,
and Model also provided their setters / adders / getters.

However, going forward, we must implement the following:

* Range detection (for auto-adjusting the ranges of X/Y/Z axes)
* Data conversion layer (for logarithmic plots, custom conversions, etc.)
* Extracting part of data (for animation features, etc.)

Considering the above,
it is reasonable to wrap all the data series by an new object: DataStore.
The DataStore stores all the data series, provides access to them.

In addition,
The DataStore performs detections, conversions, and extractions
described the above if necessary.

In this commit, we actually created the DataStore class,
and moved the data series and its accessors from Model to the DataStore.