Fix: ZxyMathDataSeries: must not have X/Y ranges explicitly
Until this commit, hasX/YMin() and haxX/YMax() methods returned true,
and getX/YMin() and getX/YMax() methods returned
the current X/Y ranges of the graph.

However, when we want to fits the X/Y ranges to the currently plotted data
(containing other data series), the ranges can not shrink smaller than
the currently configured X/Y ranges, due to the aboves.

Therefore, we fixed hasXMin() ... methods to return false.
Mathematically, the X/Y ranges of z(x,y) plotting is "unbounded",
so it should not affect to the range detection of X/Y axes.