Change: RinearnGraph3D.copyImage(boolean) to copyImage(int,boolean)
Now we can specify the type of the BufferedImage to be returned,
by the first argument.

(e.g. BufferedImage.ARGB, BufferedImage.RGB)

Also, we changed the type of the return value to more explicit type,
to BufferedImage from Image.

Probably copyImage(...) API will be mainly used for editing
the rendered graph image. The type "Image" is too abstract,
forcing the caller-side to check the actual type (= BufferedImage)
and cast to it for editing.
And if the actual type is different from BufferedImage,
the editing code on the caller-side does not work.
So we consider that the returned type shoud be restricted
to BufferedImage explicitly, for this method.