Change: API: setX/Y/ZRange methods automatically disable X/Y/Z auto-range features
On Ver.6's GUI, users can not edit X/Y/Z ranges
when the auto-range feature of each axis is enabled.
(The min/max text fields are grayout and non-editable.)

To edit the range of an axis on GUI,
the users firstly should disable the auto-range feature,
and then the text fields to edit range turn into editable.

For the consistency with the above specification on GUI, through API,
we should make it impossible to edit the range, with maintaining
the state of the auto-range feature is enabled for the axis.

Hence, considering usability,
we changed the behaviour of the setX/Y/ZRange API
to disable X/Y/Z auto-range features automatically when they are called.